Task 14, first 50 fibonacci numbers

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -11,15 +11,15 @@
 int main(void)
 {
 	int i = 3;
-	int fib_1 = 0;
-	int fib_2 = 1;
-	int fib_n = fib_1 + fib_2;
+	long int fib_1 = 0;
+	long int fib_2 = 1;
+	long int fib_n = fib_1 + fib_2;
 
-	printf("%d, %d, ", fib_1, fib_2);
+	printf("%ld, %ld, ", fib_1, fib_2);
 
 	while (i < 50)
 	{
-		printf("%d, ", fib_n);
+		printf("%ld, ", fib_n);
 		fib_1 = fib_2;
 		fib_2 = fib_n;
 		fib_n = fib_1 + fib_2;
@@ -27,7 +27,7 @@ int main(void)
 		i++;
 	}
 	fib_n = fib_1 + fib_2;
-	printf("%d\n", fib_n);
+	printf("%ld\n", fib_n);
 
 	return (0);
 }
